Rewrite update check and updater to a working state

+++ src/js/updater.js
@@ -35,7 +35,6 @@ const checkForUpdates = async () => {
 		return false;
 	} catch (e) {
 		log.write('Not updating due to error: %s', e.message);
-		console.log(e);
 		return false;
 	}
 };
@@ -44,62 +43,67 @@ const checkForUpdates = async () => {
  * Apply an outstanding update.
  */
 const applyUpdate = async () => {
-	core.block(async () => {
-		core.view.showLoadScreen('Updating, please wait...');
-		const progress = core.createProgress(2);
+	core.view.isBusy++;
+	core.view.showLoadScreen('Updating, please wait...');
 
-		const requiredFiles = [];
+	log.write('Starting update to %s...', updateManifest.guid);
 
-		// Check required files and calculate total file size...
-		let totalSize = 0;
-		const entries = Object.entries(updateManifest.contents);
-		await progress.updateWithText(1, 'Verifying local files');
-		for (let i = 0, n = entries.length; i < n; i++) {
-			const [file, meta] = entries[i];
-			const [hash, size] = meta;
+	const requiredFiles = [];
+	const entries = Object.entries(updateManifest.contents);
 
-			await progress.update(1, (i + 1) / entries.length);
-			totalSize += size;
+	let progress = core.createProgress(entries.length);
+	core.view.loadingTitle = 'Verifying local files...';
 
-			const localPath = path.join(constants.INSTALL_PATH, file);
-			const node = { file, size };
+	for (let i = 0, n = entries.length; i < n; i++) {
+		const [file, meta] = entries[i];
+		const [hash, size] = meta;
 
-			try {
-				const stats = await fsp.stat(localPath);
+		await progress.step((i + 1) + ' / ' + n);
 
-				// If the file size is different, skip hashing and just mark for update.
-				if (stats.size !== size) {
-					console.log('%d !== %d', stats.size, size);
-					requiredFiles.push(node);
-					continue;
-				}
+		const localPath = path.join(constants.INSTALL_PATH, file);
+		const node = { file, size };
 
-				// Verify local sha256 hash with remote one.
-				const localHash = await generics.getFileHash(localPath, 'sha256', 'hex');
-				if (localHash !== hash) {
-					console.log('%s !== %s', localHash, hash);
-					requiredFiles.push(node);
-					continue;
-				}
-			} catch (e) {
-				// Error thrown, likely due to file not existing.
+		try {
+			const stats = await fsp.stat(localPath);
+
+			// If the file size is different, skip hashing and just mark for update.
+			if (stats.size !== size) {
+				log.write('Marking %s for update due to size mismatch (%d != %d)', file, stats.size, size);
 				requiredFiles.push(node);
+				continue;
 			}
+
+			// Verify local sha256 hash with remote one.
+			const localHash = await generics.getFileHash(localPath, 'sha256', 'hex');
+			if (localHash !== hash) {
+				log.write('Marking %s for update due to hash mismatch (%s != %s)', file, localHash, hash);
+				requiredFiles.push(node);
+				continue;
+			}
+		} catch (e) {
+			// Error thrown, likely due to file not existing.
+			log.write('Marking %s for update due to local error: %s', file, e.message);
+			requiredFiles.push(node);
 		}
+	}
 
-		const remoteDir = util.format(core.view.config.updateURL, nw.App.manifest.flavour);
+	const downloadSize = generics.filesize(requiredFiles.map(e => e.size).reduce((total, val) => total + val));
+	log.write('%d files (%s) marked for download.', requiredFiles.length, downloadSize);
 
-		let downloadSize = 0;
-		await progress.updateWithText(2, 'Downloading updates');
-		for (const node of requiredFiles) {
-			await progress.update(2, downloadSize / totalSize);
-			await generics.downloadFile(remoteDir + node.file, path.join(constants.UPDATE.DIRECTORY, node.file));
-			downloadSize += node.size;
-		}
+	progress = core.createProgress(requiredFiles.length);
+	core.view.loadingTitle = 'Downloading updates...';
+	
+	const remoteDir = util.format(core.view.config.updateURL, nw.App.manifest.flavour);
 
-		progress.updateWithText(2, 'Restarting application', 1);
-		await launchUpdater();
-	});
+	for (let i = 0, n = requiredFiles.length; i < n; i++) {
+		const node = requiredFiles[i];
+		await progress.step((i + 1) + ' / ' + n);
+
+		await generics.downloadFile(remoteDir + node.file, path.join(constants.UPDATE.DIRECTORY, node.file));
+	}
+
+	core.view.loadingTitle = 'Restarting application...';
+	await launchUpdater();
 };
 
 /**
@@ -108,19 +112,21 @@ const applyUpdate = async () => {
 const launchUpdater = async () => {
 	// On the rare occurance that we've updated the updater, the updater
 	// cannot update the updater, so instead we update the updater here.
-	const helperApp = path.join(constants.Installpath, constants.UPDATE.HELPER);
+	const helperApp = path.join(constants.INSTALL_PATH, constants.UPDATE.HELPER);
 	const updatedApp = path.join(constants.UPDATE.DIRECTORY, constants.UPDATE.HELPER);
 
 	try {
-		// Rather than checking if an updated updater exists, just attempt
-		// to copy it regardless. It will fail if not or on permission errors.
-		await fsp.copyFile(updatedApp, helperApp);
-	} catch (e) {}
+		const updaterExists = await generics.fileExists(updatedApp);
+		if (updaterExists)
+			await fsp.copyFile(updatedApp, helperApp);
 
-	// Launch the updater application.
-	const child = cp.spawn(helperApp, [process.pid], { detached: true, stdio: 'ignore' });
-	child.unref();
-	process.exit();
+		// Launch the updater application.
+		const child = cp.spawn(helperApp, [process.pid], { detached: true, stdio: 'ignore' });
+		child.unref();
+		process.exit();
+	} catch (e) {
+		log.write('Failed to restart for update: %s', e.message);
+	}
 };
 
 module.exports = { checkForUpdates, applyUpdate };
